+++ ts/packages/defaultAgentProvider/test/data/translate-e2e.json
@@ -10,5 +10,9 @@
     {
         "request": "get my top 50 favorite songs, filter the current list to Bach pieces, and create a playlist call test",
         "action": ["player.getFavorites", "player.filterTracks", "player.createPlaylist"]
+    },
+    {
+        "request": "look up information about Oregon and write a poem using what you found",
+        "action": ["chat.lookupAndGenerateResponse", "dispatcher.pendingRequestAction"]
     }
 ]
