+++ ts/packages/dispatcher/src/context/system/handlers/configCommandHandlers.ts
@@ -627,15 +627,38 @@ const configTranslationCommandHandlers: CommandHandlerTable = {
             );
         }),
         model: new ConfigModelSetCommandHandler("translation"),
-        multi: getToggleHandlerTable(
-            "multiple action translation",
-            async (context, enable: boolean) => {
-                await changeContextConfig(
-                    { translation: { multipleActions: enable } },
-                    context,
-                );
+        multi: {
+            description: "multiple actions",
+            commands: {
+                ...getToggleCommandHandlers(
+                    "multiple action translation",
+                    async (context, enable: boolean) => {
+                        await changeContextConfig(
+                            { translation: { multiple: { enabled: enable } } },
+                            context,
+                        );
+                    },
+                ),
+                result: getToggleHandlerTable(
+                    "result id in multiple action",
+                    async (context, enable: boolean) => {
+                        await changeContextConfig(
+                            { translation: { multiple: { result: enable } } },
+                            context,
+                        );
+                    },
+                ),
+                pending: getToggleHandlerTable(
+                    "pending request in multiple action",
+                    async (context, enable: boolean) => {
+                        await changeContextConfig(
+                            { translation: { multiple: { pending: enable } } },
+                            context,
+                        );
+                    },
+                ),
             },
-        ),
+        },
         switch: {
             description: "auto switch schemas",
             commands: {
